+++ CHANGELOG.md
@@ -1,7 +1,11 @@
 # Changelog
 
 ## Unreleased
+* Added: Note add time
+* UI: Optimized the statistic card
 
+* 新增：笔记添加时间
+* UI：优化了统计卡片
 
 ## 1.1.8 2024-10-23